fix: 리뷰 상세에서 api 사용 수정
